단면 인쇄 테스트 - 프린트 함수에서 파라미터 nullable 하게 처리 - 단면 인쇄 유닛 테스트 기능 추가

+++ lib/features/core/printer/card_printer.dart
@@ -50,10 +50,14 @@ class PrinterService extends _$PrinterService {
   }
 
   Future<void> printImage({
-    required String frontImagePath,
-    required File embeddedFile,
+    required String? frontImagePath,
+    required File? embeddedFile,
   }) async {
     try {
+      if (frontImagePath == null && embeddedFile == null) {
+        throw Exception('There is nothing to print');
+      }
+
       state = const AsyncValue.loading();
       // 피더 상태 체크 추가
       logger.i('Checking feeder status...');
@@ -70,39 +74,44 @@ class PrinterService extends _$PrinterService {
       }
 
       logger.i('2. Preparing front canvas...');
-      final frontBuffer = StringBuffer();
+      StringBuffer? frontBuffer;
 
-      try {
-        await _prepareAndDrawImage(frontBuffer, frontImagePath, true);
-      } catch (e, stack) {
-        logger.i('Error in front canvas preparation: $e\nStack: $stack');
-        throw Exception('Failed to prepare front canvas: $e');
+      if (frontImagePath != null) {
+        frontBuffer = StringBuffer();
+        try {
+          await _prepareAndDrawImage(frontBuffer, frontImagePath, true);
+        } catch (e, stack) {
+          logger.i('Error in front canvas preparation: $e\nStack: $stack');
+          throw Exception('Failed to prepare front canvas: $e');
+        }
       }
 
       StringBuffer? rearBuffer;
 
-      logger.i('3. Loading and rotating rear image...');
-      final rearImage = await embeddedFile.readAsBytes();
-      final rotatedRearImage = _bindings.flipImage180(rearImage);
-      // 임시 파일로 저장
-      final temp = DateTime.now().millisecondsSinceEpoch.toString();
-      final rotatedRearPath = '${temp}_rotated.png';
-      await File(rotatedRearPath).writeAsBytes(rotatedRearImage);
-
-      try {
-        logger.i('4. Preparing rear canvas...');
-        rearBuffer = StringBuffer();
+      if (embeddedFile != null) {
+        logger.i('3. Loading and rotating rear image...');
+        final rearImage = await embeddedFile.readAsBytes();
+        final rotatedRearImage = _bindings.flipImage180(rearImage);
+        // 임시 파일로 저장
+        final temp = DateTime.now().millisecondsSinceEpoch.toString();
+        final rotatedRearPath = '${temp}_rotated.png';
+        await File(rotatedRearPath).writeAsBytes(rotatedRearImage);
 
         try {
-          await _prepareAndDrawImage(rearBuffer, rotatedRearPath, false);
-        } catch (e, stack) {
-          logger.i('Error in rear canvas preparation: $e\nStack: $stack');
-          throw Exception('Failed to prepare rear canvas: $e');
+          logger.i('4. Preparing rear canvas...');
+          rearBuffer = StringBuffer();
+
+          try {
+            await _prepareAndDrawImage(rearBuffer, rotatedRearPath, false);
+          } catch (e, stack) {
+            logger.i('Error in rear canvas preparation: $e\nStack: $stack');
+            throw Exception('Failed to prepare rear canvas: $e');
+          }
+        } finally {
+          await File(rotatedRearPath).delete().catchError((_) {
+            logger.i('Failed to delete rotated rear image');
+          });
         }
-      } finally {
-        await File(rotatedRearPath).delete().catchError((_) {
-          logger.i('Failed to delete rotated rear image');
-        });
       }
 
       logger.i('5. Injecting card...');
@@ -110,8 +119,8 @@ class PrinterService extends _$PrinterService {
 
       logger.i('6. Printing card...');
       _bindings.printCard(
-        frontImageInfo: frontBuffer.toString(),
-        backImageInfo: rearBuffer.toString(),
+        frontImageInfo: frontBuffer?.toString(),
+        backImageInfo: rearBuffer?.toString(),
       );
 
       logger.i('7. Ejecting card...');
